requirements.txt
pandas
numpy
scikit-learn
nltk
gensim
spacy
fastapi
uvicorn
openai
transformers
matplotlib
seaborn
PyPDF2
gensim
pyLDAvis
plotly
newspaper3k
lxml_html_clean
textblob
transformers
torch
shap
embedchain
mlflow
pymongo